requirements.txt
werkzeug==2.3.7
Flask
flask_sqlalchemy
flask_login
flask_wtf
WTForms
email_validator
ValidationError
flask_qrcode
pyotp
python-dotenv
MarkupSafe
from cryptography.fernet import Fernet
fernet
rsa
bcrypt
Flask-Talisman

